+++ elegy/module.py
@@ -11,8 +11,7 @@ import haiku as hk
 import jax
 import jax.numpy as jnp
 import numpy as np
-import yaml
-from tabulate import tabulate
+
 
 from elegy import types, utils
 from elegy.random import RNG
@@ -147,56 +146,6 @@ LOCAL: LocalContext = _LocalContext(
 )
 
 
-def context(
-    rng: tp.Optional[RNG] = None,
-    training: tp.Optional[bool] = None,
-    hooks: bool = False,
-    summaries: bool = False,
-) -> tp.ContextManager[LocalContext]:
-    return _context(rng=rng, training=training, hooks=hooks, summaries=summaries)
-
-
-@contextmanager
-def _context(
-    rng: tp.Optional[RNG] = None,
-    training: tp.Optional[bool] = None,
-    hooks: bool = False,
-    summaries: bool = False,
-) -> tp.Iterator[LocalContext]:
-
-    prev_rng = LOCAL.rng
-    prev_training = LOCAL.training
-    prev_losses = LOCAL.losses
-    prev_metrics = LOCAL.metrics
-    prev_summaries = LOCAL.summaries
-    prev_names = LOCAL.names
-    prev_level_names = LOCAL.level_names
-
-    LOCAL.rng = rng if rng is not None else LOCAL.rng
-    LOCAL.training = training if training is not None else LOCAL.training
-    LOCAL.losses = {} if hooks else LOCAL.losses
-    LOCAL.metrics = {} if hooks else LOCAL.metrics
-    LOCAL.summaries = [] if summaries else LOCAL.summaries
-    LOCAL.names = [] if hooks or summaries else LOCAL.names
-    LOCAL.level_names = [] if hooks or summaries else LOCAL.level_names
-
-    try:
-        yield _LocalContext(**vars(LOCAL))
-    finally:
-        # # clean
-        # if LOCAL.level_names is not None:
-        #     LOCAL.level_names.clear()
-
-        # revert
-        LOCAL.rng = prev_rng
-        LOCAL.training = prev_training
-        LOCAL.losses = prev_losses
-        LOCAL.metrics = prev_metrics
-        LOCAL.summaries = prev_summaries
-        LOCAL.names = prev_names
-        LOCAL.level_names = prev_level_names
-
-
 def construct_module(cls, *args, **kwargs) -> "Module":
     module: Module = cls.__new__(cls, *args, **kwargs)
     with instantiation_context(module):
@@ -240,9 +189,10 @@ class ModuleMeta(ABCMeta):
                 )
                 assert isinstance(module, Module)
 
-                if not isinstance(module, cls):
+                # if not isinstance(module, cls):
+                if module.__class__.__name__ != cls.__name__:
                     raise ModuleOrderError(
-                        f"Error retrieving module, expected type {cls.__name__}, got {module}. "
+                        f"Error retrieving module, expected type {cls.__name__}, got {module.__class__.__name__}. "
                         "This is probably due to control flow, you must guarantee that the same amount "
                         "of submodules will be created every time and that their order is the same."
                     )
@@ -606,159 +556,6 @@ class Module(metaclass=ModuleMeta):
                 )
             )
 
-    def summary(
-        self, x, depth: int = 2, tablefmt: str = "fancy_grid", **tablulate_kwargs
-    ):
-        """
-        Prints a summary of the network.
-
-        Arguments:
-            x: A sample of inputs to the network.
-            depth: The level number of nested level which will be showed.
-                Information about summaries from modules deeper than `depth`
-                will be aggregated together.
-            tablefmt: A string represeting the style of the table generated by
-                `tabulate`. See
-                [python-tabulate](https://github.com/astanin/python-tabulate)
-                for more options.
-            tablulate_kwargs: Additional keyword arguments passed to `tabulate`.
-                See [python-tabulate](https://github.com/astanin/python-tabulate)
-                for more options.
-        """
-        if not self.initialized:
-            self.init(x)
-
-        with context(summaries=True):
-            self(x)
-
-            summaries = get_summaries()
-
-        assert summaries is not None
-
-        def format_output(outputs) -> str:
-            file = StringIO()
-            outputs = jax.tree_map(lambda x: f"{x.shape}{{pad}}  {x.dtype}", outputs)
-            yaml.safe_dump(
-                outputs, file, default_flow_style=False, indent=2, explicit_end=False
-            )
-            return file.getvalue().replace("\n...", "")
-
-        def format_size(size):
-            return (
-                f"{size / 1e9 :,.1f} GB"
-                if size > 1e9
-                else f"{size / 1e6 :,.1f} MB"
-                if size > 1e6
-                else f"{size / 1e3 :,.1f} KB"
-                if size > 1e3
-                else f"{size:,} B"
-            )
-
-        table: tp.List = [["Inputs", format_output(x), "0", "0"]]
-
-        for module, base_name, value in summaries:
-            base_name_parts = base_name.split("/")[1:]
-            module_depth = len(base_name_parts)
-
-            if module_depth > depth:
-                continue
-
-            include_submodules = module_depth == depth
-
-            params_count = (
-                module.parameters_size(include_submodules) if module is not None else 0
-            )
-            params_size = (
-                module.parameters_bytes(include_submodules) if module is not None else 0
-            )
-            states_count = (
-                module.states_size(include_submodules) if module is not None else 0
-            )
-            states_size = (
-                module.states_bytes(include_submodules) if module is not None else 0
-            )
-            class_name = f"({module.__class__.__name__})" if module is not None else ""
-
-            base_name = "/".join(base_name_parts)
-
-            if not base_name:
-                base_name = "Outputs"
-
-            table.append(
-                [
-                    f"{base_name}{{pad}}  {class_name}",
-                    format_output(value),
-                    f"{params_count:,}{{pad}}    {format_size(params_size)}"
-                    if params_count > 0
-                    else "0",
-                    f"{states_count:,}{{pad}}    {format_size(states_size)}"
-                    if states_count > 0
-                    else "0",
-                ]
-            )
-
-        # add papdding
-        for col in range(4):
-            max_length = max(
-                len(line.split("{pad}")[0])
-                for row in table
-                for line in row[col].split("\n")
-            )
-
-            for row in table:
-                row[col] = "\n".join(
-                    line.format(
-                        pad=" " * (max_length - len(line.rstrip().split("{pad}")[0]))
-                    )
-                    for line in row[col].rstrip().split("\n")
-                )
-
-        print(
-            "\n"
-            + tabulate(
-                table,
-                headers=[
-                    "Layer",
-                    "Outputs Shape",
-                    "Trainable\nParameters",
-                    "Non-trainable\nParameters",
-                ],
-                tablefmt=tablefmt,
-                **tablulate_kwargs,
-            )
-        )
-
-        params_count = self.parameters_size()
-        params_size = self.parameters_bytes()
-        states_count = self.states_size()
-        states_size = self.states_bytes()
-        total_count = params_count + states_count
-        total_size = params_size + states_size
-
-        print(
-            tabulate(
-                [
-                    [
-                        f"Total Parameters:",
-                        f"{total_count:,}",
-                        f"{format_size(total_size)}" if total_count > 0 else "",
-                    ],
-                    [
-                        f"Trainable Parameters:",
-                        f"{params_count:,}",
-                        f"{format_size(params_size)}" if params_count > 0 else "",
-                    ],
-                    [
-                        f"Non-trainable Parameters:",
-                        f"{states_count:,}",
-                        f"{format_size(states_size)}" if states_count > 0 else "",
-                    ],
-                ],
-                tablefmt="plain",
-            )
-            + "\n"
-        )
-
 
 # -------------------------------------------------------------
 # hooks
@@ -916,6 +713,44 @@ def base_name() -> str:
 # -----------------------------------------------------------------------------
 # context managers
 # -----------------------------------------------------------------------------
+
+
+def hooks_context(
+    summaries: bool = False,
+) -> tp.ContextManager[None]:
+    return _hooks_context(summaries=summaries)
+
+
+@contextmanager
+def _hooks_context(
+    summaries: bool = False,
+) -> tp.Iterator[None]:
+
+    prev_losses = LOCAL.losses
+    prev_metrics = LOCAL.metrics
+    prev_summaries = LOCAL.summaries
+    prev_names = LOCAL.names
+    prev_level_names = LOCAL.level_names
+
+    LOCAL.losses = {}
+    LOCAL.metrics = {}
+    LOCAL.summaries = [] if summaries else LOCAL.summaries
+    LOCAL.names = []
+    LOCAL.level_names = []
+
+    try:
+        yield
+    finally:
+        # # clean
+        # if LOCAL.level_names is not None:
+        #     LOCAL.level_names.clear()
+
+        # revert
+        LOCAL.losses = prev_losses
+        LOCAL.metrics = prev_metrics
+        LOCAL.summaries = prev_summaries
+        LOCAL.names = prev_names
+        LOCAL.level_names = prev_level_names
 
 
 def rng_context(rng: RNG) -> tp.ContextManager[None]:
@@ -1337,7 +1172,7 @@ def get_unique_name(
 
 
 def to_module(f):
-    class MyModule(Module):
+    class ToModule(Module):
         def __init__(self, name: tp.Optional[str] = None):
             super().__init__(
                 name=utils.lower_snake_case(f.__name__) if name is None else name
@@ -1347,9 +1182,9 @@ def to_module(f):
         def call(self, *args, **kwargs):
             ...
 
-    MyModule.__name__ = f.__name__
+    ToModule.__name__ = f.__name__
 
-    return MyModule
+    return ToModule
 
 
 def as_initial(name):
